+++ PointingVector.cpp
@@ -14,12 +14,12 @@
 #include "PointingVector.h"
 using namespace std;
 using namespace VieVS;
-int PointingVector::nextId = 0;
+unsigned long PointingVector::nextId = 0;
 
 //PointingVector::PointingVector():VieVS_Object(nextId++), staid_{-1}, srcid_{-1}{
 //}
 
-PointingVector::PointingVector(int staid, int srcid):VieVS_Object(nextId++), staid_{staid}, srcid_{srcid}{
+PointingVector::PointingVector(unsigned long staid, unsigned long srcid):VieVS_Object(nextId++), staid_{staid}, srcid_{srcid}{
 }
 
 PointingVector::PointingVector(const PointingVector &other):
